refac: util 분리 - uiUpdate
